A few changes
Fixed plugins page with an error not showing the plugins. Still not
working correct though. See upcoming Issue that's related.

Removed Select_CSS, i'd rather it just be hardcoded in the header.tpls.
We include a default css in the static folder, users can use the static
folder for all their themes as a standardized spot or do as DJPred does
with the space theme.

Updated some issues with the Installer modules.

Fixed some warnings and errors I found in the language hook. Added
CUR_DIR because atleast on my servers, it can't find the language
directories since there's no /Language in ROOT directory. Also was a
casing issue that was fixed.